simplified provider code, tested that datasets of different modalities behave similarly concerning the target value (with the exception of Cora). Refactored model interface to remove readout and comply with new dim_input_features behavior. Removed semi-sup and sup tasks in favor of a single Standard Experiment that is usually enough
