+++ examples/DATA_CONFIGS/config_MNISTTemporal.yml
@@ -2,8 +2,8 @@ splitter:
   splits_folder: examples/DATA_SPLITS/
   class_name: mlwiz.data.splitter.Splitter
   args:
-    n_outer_folds: 3  # we use the official train/test splits of MNIST
-    n_inner_folds: 2  # not interested in doing an inner fold CV now
+    n_outer_folds: 3
+    n_inner_folds: 2
     seed: 42
     stratify: False  # no need to stratifu, the dataset is already split into training + test
     shuffle: False  # no need to shuffle, the dataset is already split into training + test
