+++ examples/DATA_CONFIGS/config_NCI1.yml
@@ -2,14 +2,14 @@ splitter:
   splits_folder: examples/DATA_SPLITS/
   class_name: mlwiz.data.splitter.Splitter
   args:
-    n_outer_folds: 3  # we use the official train/test splits of MNIST
-    n_inner_folds: 2  # not interested in doing an inner fold CV now
+    n_outer_folds: 3
+    n_inner_folds: 2
     seed: 42
-    stratify: True  # no need to stratifu, the dataset is already split into training + test
-    shuffle: True  # no need to shuffle, the dataset is already split into training + test
+    stratify: True
+    shuffle: True
     inner_val_ratio: 0.1  # let's use a portion of the training set for validation at model selection time
     outer_val_ratio: 0.1  # let's use a portion of the training set for validation at risk assessment time
-    test_ratio: 0.1  # will correspond to the last 10k samples out of 70k, see dataset.MNIST
+    test_ratio: 0.1  # not used in this case
 dataset:
   class_name: mlwiz.data.dataset.NCI1
   args:
