+++ mlwiz/model/interface.py
@@ -1,7 +1,7 @@
-from typing import Callable, Tuple, Optional, List
+from typing import Callable, Tuple, Optional, List, Union
 
 import torch
-from torch_geometric.data import Batch
+import torch_geometric
 
 
 class ModelInterface(torch.nn.Module):
@@ -9,62 +9,7 @@ class ModelInterface(torch.nn.Module):
     Provides the signature for any main model to be trained under MLWiz
 
     Args:
-        dim_input_features (int): dimension of node features
-            (according to the :class:`~mlwiz.data.dataset.DatasetInterface`
-            property)
-        dim_edge_features (int): dimension of edge features
-            (according to the :class:`~mlwiz.data.dataset.DatasetInterface`
-            property)
-        dim_target (int): dimension of the target
-            (according to the :class:`~mlwiz.data.dataset.DatasetInterface`
-            property)
-        readout_class (Callable[...,:class:`torch.nn.Module`]):
-            class of the module implementing the readout. This is optional,
-            but useful to put different readouts to try in the config file
-        config (dict): config dictionary containing all the necessary
-            hyper-parameters plus additional information (if needed)
-    """
-
-    def __init__(
-        self,
-        dim_input_features: int,
-        dim_edge_features: int,
-        dim_target: int,
-        readout_class: Callable[..., torch.nn.Module],
-        config: dict,
-    ):
-        super().__init__()
-        self.dim_input_features = dim_input_features
-        self.dim_edge_features = dim_edge_features
-        self.dim_target = dim_target
-        self.readout_class = readout_class
-        self.config = config
-
-    def forward(
-        self, data: Batch
-    ) -> Tuple[torch.Tensor, Optional[torch.Tensor], Optional[List[object]]]:
-        r"""
-        Performs a forward pass over a batch of graphs
-
-        Args:
-            data (:class:`torch_geometric.data.Batch`): a batch of graphs
-
-        Returns:
-            a tuple (model's output, [optional] node embeddings,
-            [optional] additional outputs
-        """
-        raise NotImplementedError("You need to implement this method!")
-
-
-class ReadoutInterface(torch.nn.Module):
-    r"""
-    Provides the signature for any readout to be trained under MLWiz
-
-    Args:
-        dim_input_features (int): dimension of node features
-            (according to the :class:`~mlwiz.data.dataset.DatasetInterface`
-            property)
-        dim_edge_features (int): dimension of edge features
+        dim_input_features (Union[int, Tuple[int]]): dimension of node features
             (according to the :class:`~mlwiz.data.dataset.DatasetInterface`
             property)
         dim_target (int): dimension of the target
@@ -76,28 +21,23 @@ class ReadoutInterface(torch.nn.Module):
 
     def __init__(
         self,
-        dim_input_features: int,
-        dim_edge_features: int,
+        dim_input_features: Union[int, Tuple[int]],
         dim_target: int,
         config: dict,
     ):
         super().__init__()
         self.dim_input_features = dim_input_features
-        self.dim_edge_features = dim_edge_features
         self.dim_target = dim_target
         self.config = config
 
     def forward(
-        self, node_embeddings: torch.tensor, batch: torch.Tensor, **kwargs
+        self, data: Union[torch.Tensor, torch_geometric.data.Batch]
     ) -> Tuple[torch.Tensor, Optional[torch.Tensor], Optional[List[object]]]:
         r"""
         Performs a forward pass over a batch of graphs
 
         Args:
-            node_embeddings (:class:`torch_geometric.data.Batch`):
-                the node embeddings
-            batch (:class:`torch.Tensor`): the usual ``batch`` object of PyG
-            kwargs (dict): additional and optional arguments
+            data: a batch of samples
 
         Returns:
             a tuple (model's output, [optional] node embeddings,
